Move packet processing example as a new, simpler example: ipv6-hooks
