+++ examples/ipv6-hooks/README.md
@@ -1,0 +1,2 @@
+A simple example of how to set up IPv6 packet hooks with netstack_ip_packet_processor_add.
+You can run this example as native and then ping the node to see printouts from the hooks, at every packet input/output.
